Disable dock/pin buttons for maximized widgets.
