+++ ui/appui-react/src/appui-react/layout/widget/Buttons.tsx
@@ -8,21 +8,14 @@
 
 import "./Buttons.scss";
 import * as React from "react";
-import { assert } from "@itwin/core-bentley";
-import { useLayout } from "../base/LayoutStore";
-import {
-  isHorizontalPanelSide,
-  PanelSideContext,
-} from "../widget-panels/Panel";
-import { Dock } from "./Dock";
-import { PinToggle } from "./PinToggle";
+import { Dock, useDock } from "./Dock";
+import { PinToggle, usePinToggle } from "./PinToggle";
 import { PopoutToggle, usePopoutToggle } from "./PopoutToggle";
 import {
   PreviewHorizontalPanelAlignButton,
   useHorizontalPanelAlignButton,
 } from "./PreviewHorizontalPanelAlign";
 import { SendBack, useSendBack } from "./SendBack";
-import { WidgetIdContext } from "./Widget";
 import {
   MoreButton,
   useDropdownFeatures,
@@ -65,15 +58,12 @@ export function TabBarButtons() {
 
 /** @internal */
 export function useWidgetFeatures() {
-  const isToolSettings = useIsToolSettingsTab();
-  const isMainPanelWidget = useIsMainPanelWidget();
-
   const popoutToggle = usePopoutToggle();
   const maximizeToggle = useMaximizeToggle();
   const sendBack = useSendBack();
-  const dock = isToolSettings;
+  const dock = useDock();
   const horizontalPanelAlignButton = useHorizontalPanelAlignButton();
-  const pinToggle = isMainPanelWidget;
+  const pinToggle = usePinToggle();
 
   return [
     ...(popoutToggle ? (["popout"] as const) : []),
@@ -83,19 +73,4 @@ export function useWidgetFeatures() {
     ...(horizontalPanelAlignButton ? (["horizontalAlign"] as const) : []),
     ...(pinToggle ? (["pin"] as const) : []),
   ];
-}
-
-/** @internal */
-export function useIsMainPanelWidget() {
-  const side = React.useContext(PanelSideContext);
-  const widgetId = React.useContext(WidgetIdContext);
-  return useLayout((state) => {
-    const widgets = side ? state.panels[side].widgets : undefined;
-    if (!widgets) return false;
-    assert(!!side);
-    const mainWidget = isHorizontalPanelSide(side)
-      ? widgets[widgets.length - 1]
-      : widgets[0];
-    return mainWidget === widgetId;
-  });
 }
