+++ ui/appui-react/src/appui-react/layout/widget/PinToggle.tsx
@@ -13,6 +13,8 @@ import { NineZoneDispatchContext, useLabel } from "../base/NineZone";
 import { PanelSideContext } from "../widget-panels/Panel";
 import { useLayout } from "../base/LayoutStore";
 import { ActionButton } from "../../preview/widget-action-dropdown/Button";
+import { useIsMainPanelWidget } from "./useIsMainPanelWidget";
+import { useMaximizedWidget } from "../../preview/enable-maximized-widget/useMaximizedWidget";
 
 /** @internal */
 export function PinToggle() {
@@ -35,4 +37,11 @@ export function PinToggle() {
       }}
     />
   );
+}
+
+/** @internal */
+export function usePinToggle() {
+  const isMainPanelWidget = useIsMainPanelWidget();
+  const maximizedWidget = useMaximizedWidget();
+  return isMainPanelWidget && !maximizedWidget;
 }
